Refactoring testWorkings
